requirements.txt
google-generativeai
python-dotenv
pdf2image
pillow
reportlab
pyinstaller
